[WFCORE-5904] Fix typo in some Handler filenames.
