+++ domain-management/src/main/java/org/jboss/as/domain/management/access/AccessAuthorizationUseIdentityRolesWriteAttributeHandler.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2013, Red Hat, Inc., and individual contributors
+ * Copyright 2016, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -19,35 +19,45 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.domain.management.access;
 
-import java.util.Locale;
+package org.jboss.as.domain.management.access;
 
 import org.jboss.as.controller.AbstractWriteAttributeHandler;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.access.management.WritableAuthorizerConfiguration;
+import org.jboss.as.controller.registry.Resource;
 import org.jboss.dmr.ModelNode;
 
 /**
- * A {@link OperationStepHandler} for writing the include-all attribute.
+ * An {@link org.jboss.as.controller.OperationStepHandler} handling write updates to the 'use-identity-roles' attribute.
  *
  * @author <a href="mailto:darran.lofthouse@jboss.com">Darran Lofthouse</a>
  */
-public class RoleIncludeAllWriteAttributeHander extends AbstractWriteAttributeHandler<Void> {
+class AccessAuthorizationUseIdentityRolesWriteAttributeHandler extends AbstractWriteAttributeHandler<Void> {
 
-    private final WritableAuthorizerConfiguration authorizerConfiguration;
+    private final WritableAuthorizerConfiguration writableConfiguration;
 
-    RoleIncludeAllWriteAttributeHander(final WritableAuthorizerConfiguration authorizerConfiguration) {
-        super(RoleMappingResourceDefinition.INCLUDE_ALL);
-        this.authorizerConfiguration = authorizerConfiguration;
+    AccessAuthorizationUseIdentityRolesWriteAttributeHandler(WritableAuthorizerConfiguration writableConfiguration) {
+        super(AccessAuthorizationResourceDefinition.USE_IDENTITY_ROLES);
+        this.writableConfiguration = writableConfiguration;
+    }
+
+    @Override
+    protected void finishModelStage(OperationContext context, ModelNode operation, String attributeName, ModelNode newValue,
+            ModelNode oldValue, Resource model) throws OperationFailedException {
+        boolean useIdentityRoles = newValue.asBoolean();
+         if (useIdentityRoles == false) {
+             /*
+              * As we are no longer using identity roles we may need another role mapping strategy.
+              */
+             RbacSanityCheckOperation.addOperation(context);
+         }
     }
 
     @Override
     protected boolean requiresRuntime(OperationContext context) {
-        // During boot add would have taken care of this, otherwise
-        // all process types need a runtime update.
-        return context.isBooting() == false;
+        return true;
     }
 
     @Override
@@ -55,20 +65,27 @@ public class RoleIncludeAllWriteAttributeHander extends AbstractWriteAttributeHa
             ModelNode resolvedValue, ModelNode currentValue,
             org.jboss.as.controller.AbstractWriteAttributeHandler.HandbackHolder<Void> handbackHolder)
             throws OperationFailedException {
+        if (!resolvedValue.equals(currentValue)) {
+            if (!context.isBooting()) {
+                return true;
+            }
+            updateAuthorizer(resolvedValue, writableConfiguration);
+        }
 
-        String roleName = RoleMappingResourceDefinition.getRoleName(operation);
-        authorizerConfiguration.setRoleMappingIncludeAll(roleName.toUpperCase(Locale.ENGLISH), resolvedValue.asBoolean());
         return false;
     }
 
     @Override
     protected void revertUpdateToRuntime(OperationContext context, ModelNode operation, String attributeName,
             ModelNode valueToRestore, ModelNode valueToRevert, Void handback) throws OperationFailedException {
-        String roleName = RoleMappingResourceDefinition.getRoleName(operation);
-        authorizerConfiguration.setRoleMappingIncludeAll(roleName, valueToRestore.asBoolean());
+        updateAuthorizer(valueToRestore, writableConfiguration);
     }
 
+    static void updateAuthorizer(final ModelNode value, final WritableAuthorizerConfiguration writableConfiguration) {
+        ModelNode resolvedValue = value.isDefined() ? value : AccessAuthorizationResourceDefinition.USE_IDENTITY_ROLES.getDefaultValue();
+        boolean useIdentityRoles = resolvedValue.asBoolean();
 
-
+        writableConfiguration.setUseIdentityRoles(useIdentityRoles);
+    }
 
 }
