+++ domain-management/src/main/java/org/jboss/as/domain/management/access/RoleIncludeAllWriteAttributeHandler.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2016, Red Hat, Inc., and individual contributors
+ * Copyright 2013, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -19,45 +19,35 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-
 package org.jboss.as.domain.management.access;
+
+import java.util.Locale;
 
 import org.jboss.as.controller.AbstractWriteAttributeHandler;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.access.management.WritableAuthorizerConfiguration;
-import org.jboss.as.controller.registry.Resource;
 import org.jboss.dmr.ModelNode;
 
 /**
- * An {@link org.jboss.as.controller.OperationStepHandler} handling write updates to the 'use-identity-roles' attribute.
+ * An {@link org.jboss.as.controller.OperationStepHandler} for writing the include-all attribute.
  *
  * @author <a href="mailto:darran.lofthouse@jboss.com">Darran Lofthouse</a>
  */
-class AccessAuthorizationUseIdentityRolesWriteAttributeHander extends AbstractWriteAttributeHandler<Void> {
+public class RoleIncludeAllWriteAttributeHandler extends AbstractWriteAttributeHandler<Void> {
 
-    private final WritableAuthorizerConfiguration writableConfiguration;
+    private final WritableAuthorizerConfiguration authorizerConfiguration;
 
-    AccessAuthorizationUseIdentityRolesWriteAttributeHander(WritableAuthorizerConfiguration writableConfiguration) {
-        super(AccessAuthorizationResourceDefinition.USE_IDENTITY_ROLES);
-        this.writableConfiguration = writableConfiguration;
-    }
-
-    @Override
-    protected void finishModelStage(OperationContext context, ModelNode operation, String attributeName, ModelNode newValue,
-            ModelNode oldValue, Resource model) throws OperationFailedException {
-        boolean useIdentityRoles = newValue.asBoolean();
-         if (useIdentityRoles == false) {
-             /*
-              * As we are no longer using identity roles we may need another role mapping strategy.
-              */
-             RbacSanityCheckOperation.addOperation(context);
-         }
+    RoleIncludeAllWriteAttributeHandler(final WritableAuthorizerConfiguration authorizerConfiguration) {
+        super(RoleMappingResourceDefinition.INCLUDE_ALL);
+        this.authorizerConfiguration = authorizerConfiguration;
     }
 
     @Override
     protected boolean requiresRuntime(OperationContext context) {
-        return true;
+        // During boot add would have taken care of this, otherwise
+        // all process types need a runtime update.
+        return context.isBooting() == false;
     }
 
     @Override
@@ -65,27 +55,20 @@ class AccessAuthorizationUseIdentityRolesWriteAttributeHander extends AbstractWr
             ModelNode resolvedValue, ModelNode currentValue,
             org.jboss.as.controller.AbstractWriteAttributeHandler.HandbackHolder<Void> handbackHolder)
             throws OperationFailedException {
-        if (!resolvedValue.equals(currentValue)) {
-            if (!context.isBooting()) {
-                return true;
-            }
-            updateAuthorizer(resolvedValue, writableConfiguration);
-        }
 
+        String roleName = RoleMappingResourceDefinition.getRoleName(operation);
+        authorizerConfiguration.setRoleMappingIncludeAll(roleName.toUpperCase(Locale.ENGLISH), resolvedValue.asBoolean());
         return false;
     }
 
     @Override
     protected void revertUpdateToRuntime(OperationContext context, ModelNode operation, String attributeName,
             ModelNode valueToRestore, ModelNode valueToRevert, Void handback) throws OperationFailedException {
-        updateAuthorizer(valueToRestore, writableConfiguration);
+        String roleName = RoleMappingResourceDefinition.getRoleName(operation);
+        authorizerConfiguration.setRoleMappingIncludeAll(roleName, valueToRestore.asBoolean());
     }
 
-    static void updateAuthorizer(final ModelNode value, final WritableAuthorizerConfiguration writableConfiguration) {
-        ModelNode resolvedValue = value.isDefined() ? value : AccessAuthorizationResourceDefinition.USE_IDENTITY_ROLES.getDefaultValue();
-        boolean useIdentityRoles = resolvedValue.asBoolean();
 
-        writableConfiguration.setUseIdentityRoles(useIdentityRoles);
-    }
+
 
 }
